move push func to separate file and adding tidwall/gjson to credits
