+++ system/api/handlers.go
@@ -11,7 +11,6 @@ import (
 	"github.com/ponzu-cms/ponzu/system/api/analytics"
 	"github.com/ponzu-cms/ponzu/system/db"
 	"github.com/ponzu-cms/ponzu/system/item"
-	"github.com/tidwall/gjson"
 )
 
 func typesHandler(res http.ResponseWriter, req *http.Request) {
@@ -99,7 +98,8 @@ func contentHandler(res http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	if pt, ok := item.Types[t]; !ok {
+	pt, ok := item.Types[t]
+	if !ok {
 		res.WriteHeader(http.StatusNotFound)
 		return
 	}
@@ -181,33 +181,6 @@ func toJSON(data []string) ([]byte, error) {
 	}
 
 	return buf.Bytes(), nil
-}
-
-func push(res http.ResponseWriter, pt func() interface{}, data []byte) {
-	// Push(target string, opts *PushOptions) error
-	if pusher, ok := res.(http.Pusher); ok {
-		if p, ok := pt().(item.Pushable); ok {
-			// get fields to pull values from data
-			fields := p.Push()
-
-			// parse values from data to push
-			values := gjson.GetManyBytes(data, fields...)
-
-			// push all values from Pushable items' fields
-			for i := range values {
-				val := values[i]
-				val.ForEach(func(k, v gjson.Result) bool {
-					err := pusher.Push(v.String(), nil)
-					if err != nil {
-						log.Println("Error during Push of value:", v.String())
-					}
-
-					return true
-				})
-			}
-		}
-	}
-
 }
 
 // sendData() should be used any time you want to communicate
